requirements.txt
emails==0.6
Flask==2.2.2
gcloud==0.18.3
googleapis-common-protos==1.57.0
gunicorn==20.1.0
Jinja2==3.1.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.2.0
Pyrebase4==4.5.0
requests==2.27.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
secure-smtplib==0.1.1
sklearn==0.0.post1
virtualenv==20.16.6
